# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {},
        "license": {}
    },
    "paths": {
        "/admin": {
            "get": {
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Get own details",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/swagger.adminGetSelf200"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Edit own details",
                "parameters": [
                    {
                        "description": "JSON request body\u003cbr\u003e\u003cbr\u003e* All fields are optional.\u003cbr\u003e** If field left blank it will not be updated.",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/swagger.adminEditSelfBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/swagger.adminEditAdmin200"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/swagger.adminEditAdmin409"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/swagger.adminEditAdmin422"
                        }
                    }
                }
            }
        },
...
        "/clients": {
            "get": {
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Get all clients",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/swagger.adminGetAllClients200"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/swagger.adminGetClientByID404"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Create a new client",
                "parameters": [
                    {
                        "description": "JSON request body\u003cbr\u003e\u003cbr\u003e* Phone is optional\u003cbr\u003e** Notes is optional",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/swagger.adminCreateClientBody"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/swagger.adminCreateClient201"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/swagger.adminCreateClient409"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/swagger.adminCreateClient422"
                        }
                    }
                }
            }
        },
        "/clients/{id}": {
            "get": {
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Get client by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Admin ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/swagger.adminGetClientByID200"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/swagger.adminGetClientByID404"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/swagger.adminGetClientByID422"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Edit a client",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Client ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "JSON request body\u003cbr\u003e\u003cbr\u003e* All fields are optional.\u003cbr\u003e** If field left blank it will not be updated.",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/swagger.adminEditClientBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/swagger.adminEditClient200"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/swagger.adminEditClient404"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/swagger.adminEditClient409"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/swagger.adminEditClient422"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin: SuperAdmin"
                ],
                "summary": "\"Soft\" delete a client",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Client ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/swagger.adminDeleteClient200"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/swagger.adminDeleteClient404"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/swagger.adminDeleteClient422"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "swagger.adminCreateAdmin1": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "admin@test.com"
                },
                "first_name": {
                    "type": "string",
                    "example": "John"
                },
                "id": {
                    "type": "integer",
                    "example": 5
                },
                "last_name": {
                    "type": "string",
                    "example": "Doe"
                },
                "role": {
                    "type": "string",
                    "example": "admin"
                }
            }
        },
...
        "swagger.adminCreateClient1": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "client@test.com"
                },
                "first_name": {
                    "type": "string",
                    "example": "John"
                },
                "id": {
                    "type": "integer",
                    "example": 5
                },
                "last_name": {
                    "type": "string",
                    "example": "Doe"
                },
                "notes": {
                    "type": "string",
                    "example": "Random text about client"
                },
                "phone": {
                    "type": "string",
                    "example": "(702) 867-5309"
                },
                "status": {
                    "type": "string",
                    "example": "active"
                }
            }
        },
        "swagger.adminCreateClient201": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object",
                    "$ref": "#/definitions/swagger.adminCreateClient1"
                },
                "message": {
                    "type": "string",
                    "example": "Created new client"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "swagger.adminCreateClient409": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Client with email already exists"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "swagger.adminCreateClient422": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "First name is required"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "swagger.adminCreateClientBody": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "client@test.com"
                },
                "first_name": {
                    "type": "string",
                    "example": "John"
                },
                "last_name": {
                    "type": "string",
                    "example": "Doe"
                },
                "notes": {
                    "type": "string",
                    "example": "Random text about client"
                },
                "password": {
                    "type": "string",
                    "example": "password"
                },
                "phone": {
                    "type": "string",
                    "example": "(702) 867-5309"
                }
            }
        },
...
        "swagger.adminDeleteClient200": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Deleted client with ID of 13"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "swagger.adminDeleteClient404": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Active client not found"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "swagger.adminDeleteClient422": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "ID in invalid format"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "swagger.adminEditAdmin200": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object",
                    "$ref": "#/definitions/swagger.adminCreateAdmin1"
                },
                "message": {
                    "type": "string",
                    "example": "Updated admin"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "swagger.adminEditAdmin409": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Email address already in use"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "swagger.adminEditAdmin422": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Email is in an invalid format"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
...
        "swagger.adminEditClient200": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object",
                    "$ref": "#/definitions/swagger.adminCreateClient1"
                },
                "message": {
                    "type": "string",
                    "example": "Updated client"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "swagger.adminEditClient404": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Active client not found"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "swagger.adminEditClient409": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Email address already in use"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "swagger.adminEditClient422": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Email is in an invalid format"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "swagger.adminEditClientBody": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "admin@test.com"
                },
                "first_name": {
                    "type": "string",
                    "example": "John"
                },
                "last_name": {
                    "type": "string",
                    "example": "Doe"
                },
                "notes": {
                    "type": "string",
                    "example": "This is a test note full of random test that is related to this client."
                },
                "password": {
                    "type": "string",
                    "example": "password"
                },
                "phone": {
                    "type": "string",
                    "example": "(702) 867-5309"
                },
                "status": {
                    "type": "string",
                    "example": "active"
                }
            }
        },
        "swagger.adminEditSelfBody": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "admin@test.com"
                },
                "first_name": {
                    "type": "string",
                    "example": "John"
                },
                "last_name": {
                    "type": "string",
                    "example": "Doe"
                },
                "password": {
                    "type": "string",
                    "example": "password"
                }
            }
        },
        "swagger.adminGetAll1": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": "2020-08-07T20:12:54Z"
                },
                "email": {
                    "type": "string",
                    "example": "admin@test.com"
                },
                "first_name": {
                    "type": "string",
                    "example": "John"
                },
                "id": {
                    "type": "integer",
                    "example": 5
                },
                "last_name": {
                    "type": "string",
                    "example": "Doe"
                },
                "role": {
                    "type": "string",
                    "example": "admin"
                },
                "status": {
                    "type": "string",
                    "example": "active"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2020-08-07T20:12:54Z"
                }
            }
        },
...
        "swagger.adminGetAllClients1": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": "2020-08-07T20:12:54Z"
                },
                "email": {
                    "type": "string",
                    "example": "admin@test.com"
                },
                "first_name": {
                    "type": "string",
                    "example": "John"
                },
                "id": {
                    "type": "integer",
                    "example": 5
                },
                "last_name": {
                    "type": "string",
                    "example": "Doe"
                },
                "notes": {
                    "type": "string",
                    "example": "This is a test note full of random test that is related to this client."
                },
                "phone": {
                    "type": "string",
                    "example": "(702) 867-5309"
                },
                "status": {
                    "type": "string",
                    "example": "active"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2020-08-07T20:12:54Z"
                }
            }
        },
        "swagger.adminGetAllClients200": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/swagger.adminGetAll1"
                    }
                },
                "message": {
                    "type": "string",
                    "example": "All clients"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                },
                "total_results": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
...
        "swagger.adminGetClientByID200": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object",
                    "$ref": "#/definitions/swagger.adminGetAllClients1"
                },
                "message": {
                    "type": "string",
                    "example": "Returned client"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "swagger.adminGetClientByID404": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Client not found"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "swagger.adminGetClientByID422": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "ID in invalid format"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "swagger.adminGetSelf200": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/swagger.adminGetAll1"
                    }
                },
                "message": {
                    "type": "string",
                    "example": "Returned admin"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
    },
    "securityDefinitions": {
        "bearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}